api/requirements.txt
flask==3.0.3
Pillow==10.3.0
python-slugify==8.0.4
